pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
authors = [
  {name = "Moritz Breurather", email = "breurather@ift.at"},
  {name = "René Schwaiger", email = "rene.schwaiger@ift.at"},
]

dependencies = [
  "icotronic>=4,<5",
  "netaddr>=1.3.0",
  "pyee>=13.0.0",
]
description = """Stateful API for the ICOtronic system intended for usage in \
              a Web API"""
keywords = [
  "smart-tool",
  "smh",
  "stu",
  "sth",
  "tool-holder",
]
name = "icostate"
readme = "ReadMe.md"
requires-python = ">=3.10,<4"
version = "0.13"

[project.optional-dependencies]
dev = [
  "black>=25.1.0",
  "coverage>=7.9.2",
  "Flake8-pyproject>=1.2.3",
  "furo>=2024.8.6",
  "mypy>=1.16.1",
  "myst-parser>=4.0.1",
  "pydoclint[flake8]>=0.6.6",
  "pylint>=3.3.7",
  "sphinx-autodoc-typehints>=3",
  "sphinxcontrib-mermaid>=1.0.0",
  "sphinx-pyproject>=0.3",
  "sphinx_toggleprompt>=0.6.0",
]
test = [
  "pytest-asyncio>=1.0.0",
  "pytest-sphinx>=0.6.3",
]

[project.urls]
Documentation = "https://icostate.readthedocs.io"
Source = "https://github.com/mytoolit/ICOstate"

[tool.black]
line-length = 79
preview = true
unstable = true
enable-unstable-feature = ["string_processing"]

[tool.flake8]
ignore = [
  # Whitespace before colon
  "E203",
  # Module level import not at top of file
  "E402",
  # Line break before binary operator
  "W503",
  # Line break after binary operator
  "W504"
]
per-file-ignores = "__init__.py:F401,conf.py:F821"

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true

[tool.pytest.ini_options]
# - Use doctests
# - Verbose output
addopts = "--doctest-modules -v"
minversion = "6.0"
asyncio_mode = "auto"
